backend/requirements.txt
Django>=5.0
django-cors-headers>=4.0
djangorestframework>=3.15

gunicorn
opencv-python
pyserial